requirements.txt
numpy==1.26.4
Pillow==10.4.0
deap==1.4.1
matplotlib==3.9.2
scipy==1.14.1
opencv-python==4.10.0.84
